LPS-55928 moved methods to RelatedEntryIndexer
