+++ portal-impl/src/com/liferay/portlet/documentlibrary/util/DLFileEntryIndexer.java
@@ -25,6 +25,7 @@ import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.repository.model.FileEntry;
 import com.liferay.portal.kernel.repository.model.FileVersion;
 import com.liferay.portal.kernel.search.BaseIndexer;
+import com.liferay.portal.kernel.search.BaseRelatedEntryIndexer;
 import com.liferay.portal.kernel.search.BooleanClauseOccur;
 import com.liferay.portal.kernel.search.BooleanQuery;
 import com.liferay.portal.kernel.search.BooleanQueryFactoryUtil;
@@ -35,6 +36,7 @@ import com.liferay.portal.kernel.search.DocumentImpl;
 import com.liferay.portal.kernel.search.Field;
 import com.liferay.portal.kernel.search.Indexer;
 import com.liferay.portal.kernel.search.IndexerRegistryUtil;
+import com.liferay.portal.kernel.search.RelatedEntryIndexer;
 import com.liferay.portal.kernel.search.SearchContext;
 import com.liferay.portal.kernel.search.SearchEngineUtil;
 import com.liferay.portal.kernel.search.SearchException;
@@ -98,7 +100,7 @@ import javax.portlet.PortletResponse;
  */
 @OSGiBeanProperties
 public class DLFileEntryIndexer
-	extends BaseIndexer implements DDMStructureIndexer {
+	extends BaseIndexer implements DDMStructureIndexer, RelatedEntryIndexer {
 
 	public static final String CLASS_NAME = DLFileEntry.class.getName();
 
@@ -109,6 +111,14 @@ public class DLFileEntryIndexer
 			Field.MODIFIED_DATE, Field.SCOPE_GROUP_ID, Field.TITLE, Field.UID);
 		setFilterSearch(true);
 		setPermissionAware(true);
+	}
+
+	@Override
+	public void addRelatedClassNames(
+			BooleanQuery contextQuery, SearchContext searchContext)
+		throws Exception {
+
+		_relatedEntryIndexer.addRelatedClassNames(contextQuery, searchContext);
 	}
 
 	@Override
@@ -447,8 +457,13 @@ public class DLFileEntryIndexer
 				Indexer indexer = IndexerRegistryUtil.getIndexer(
 					dlFileEntry.getClassName());
 
-				if (indexer != null) {
-					indexer.addRelatedEntryFields(
+				if ((indexer != null) &&
+					(indexer instanceof RelatedEntryIndexer)) {
+
+					RelatedEntryIndexer relatedEntryIndexer =
+						(RelatedEntryIndexer)indexer;
+
+					relatedEntryIndexer.addRelatedEntryFields(
 						document, new LiferayFileEntry(dlFileEntry));
 
 					DocumentHelper documentHelper = new DocumentHelper(
@@ -679,5 +694,8 @@ public class DLFileEntryIndexer
 
 	private static final Log _log = LogFactoryUtil.getLog(
 		DLFileEntryIndexer.class);
+
+	private final RelatedEntryIndexer _relatedEntryIndexer =
+		new BaseRelatedEntryIndexer();
 
 }
