+++ portal-service/src/com/liferay/portal/kernel/search/RelatedEntryIndexer.java
@@ -1,0 +1,31 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.kernel.search;
+
+/**
+ * @author Michael C. Han
+ */
+public interface RelatedEntryIndexer {
+
+	public void addRelatedClassNames(
+			BooleanQuery contextQuery, SearchContext searchContext)
+		throws Exception;
+
+	public void addRelatedEntryFields(Document document, Object obj)
+		throws Exception;
+
+	public void updateFullQuery(SearchContext searchContext);
+
+}
